codepaths.py: a way to dump leaves of a function
This is meant to be used with inject-error.py.  Basically you have a
function you want to test errors with, you run codepaths.py on that
function to get it to dump all of the possible originating functions for
that function, and then systematically call inject-error.py with the
originating function as the trigger and the function you want to fail as
the fail target.  Then you just walk through the list until something
blows up.

Signed-off-by: Josef Bacik <jbacik@fb.com>