kernelparse: basic regex library for kernel parsing
Most c parsers for python use all this complicated shit that doesn't
work with the kernel because we use gnu extensions.  Honestly we really
just need callers and callee's, and that can be done with regex.  This
is a braindead library to let me do simple analysis for finding
codepaths for error injection and such.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>