enospc-stress: watch the btrfs_fail_tickets tracepoint
There are a few cases where we'll get an "early" enospc where it really
isn't "early".  This happens sometimes because generally we fill up the
entire disk with data, and then have to wait for the metadata filler to
go.  This means we'll get basically no metadata chunks allocated even
tho we're really close to the global rsv.  However if we unlink
everybody we'll free a bunch of block groups, and we could potentially
allocate a metadata chunk during this time, throwing off our ability to
fill up with data.

To deal with this instead watch the btrfs_fail_tickets tracepoint to
determine if this was really an early ENOSPC.  This will allow us to
have these kind of transient failures without claiming that the test
failed.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>